# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🎶 MusicBoxAPI",
        "description": "API RESTful pour la gestion des artistes, albums et chansons.\n *                   Cette API permet de gérer les ressources principales :\n *                   - Artistes\n *                   - Albums\n *                   - Chansons\n *                   Elle inclut des fonctionnalités de recherche, pagination et filtres.",
        "contact": {
            "name": "MusicBox API Team",
            "email": "support@musicboxapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Lister tous les albums",
                "operationId": "735609843dcfb89ee4e12e340eddca25",
                "responses": {
                    "200": {
                        "description": "Liste des albums paginée"
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Créer un nouvel album",
                "operationId": "2252aa8fd63b7d2fa1203c1735943d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Marshall Mathers LP"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album créé avec succès"
                    }
                }
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Afficher un album par ID",
                "operationId": "71e9c4179582355e89419e6f6994c418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album trouvé"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Mettre à jour un album",
                "operationId": "1300ec85cc590b1ae1468c9bf8db122d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Encore"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2004
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album mis à jour avec succès"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Supprimer un album",
                "operationId": "cabfc7c152115607d310827d46a82d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Album supprimé avec succès"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            }
        },
        "/api/albums/{id}/songs": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Lister les chansons d'un album",
                "operationId": "2d86f07a20921b0701894cdb6aedcc80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des chansons de l'album"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}